Most of the circular dependencies removed
