+++ src/ducks/Chart/utils.js
@@ -1,4 +1,5 @@
 import { Metric } from '../dataHub/metrics'
+import { MirroredMetric } from '../dataHub/metrics/mirrored'
 import {
   getDateFormats,
   getTimeFormats,
@@ -101,4 +102,25 @@ export function findPointByDate (points, target) {
       index -= 1
     }
   }
+}
+
+export function extractMirrorMetricsDomainGroups (domainGroups) {
+  if (!domainGroups) return
+
+  const mirroredGroups = []
+
+  const { length } = domainGroups
+  for (let i = 0; i < length; i++) {
+    const group = domainGroups[i]
+    for (let y = group.length; y > 0; y--) {
+      const metricKey = group[y]
+
+      if (MirroredMetric[metricKey]) {
+        mirroredGroups.push(group)
+        break
+      }
+    }
+  }
+
+  return mirroredGroups
 }
