+++ src/ducks/Signals/signalFormManager/sharedForm/SharedTriggerForm.js
@@ -3,7 +3,6 @@ import { connect } from 'react-redux'
 import Button from '@santiment-network/ui/Button'
 import { couldShowChart } from '../../utils/utils'
 import SignalPreview from '../../chart/preview/SignalPreview'
-import SignalCard from '../../../../components/SignalCard/card/SignalCard'
 import NoSignalPreview from '../../chart/preview/NoSignalPreview'
 import { DesktopOnly, MobileOnly } from '../../../../components/Responsive'
 import CopySignal from '../../../../components/SignalCard/controls/CopySignal'
@@ -17,7 +16,8 @@ const SharedTriggerForm = ({
   onCreate,
   settings,
   originalTrigger,
-  isAuthor
+  isAuthor,
+  SignalCard
 }) => {
   const { metric } = settings
