+++ src/ducks/Studio/Chart/index.js
@@ -12,7 +12,6 @@ import ContextMenu from './ContextMenu'
 import ChartFullscreenBtn from './Fullscreen'
 import Insights from './Insights'
 import Compare from '../Compare'
-import { extractMirrorMetricsDomainGroups } from '../utils'
 import { useMetricColor } from '../Widget/ChartWidgetColorProvider'
 import { useAllTimeData } from '../timeseries/hooks'
 import Chart from '../../Chart'
@@ -20,6 +19,7 @@ import Signals from '../../Chart/Signals'
 import { useMetricCategories } from '../../Chart/Synchronizer'
 import { useDomainGroups, useAxesMetricsKey } from '../../Chart/hooks'
 import { useHighlightMetricColor } from '../../Chart/colors'
+import { extractMirrorMetricsDomainGroups } from '../../Chart/utils'
 import { useUser } from '../../../stores/user'
 import styles from './index.module.scss'
